+++ pkgs/artisanal_widgets/example/artisanal_app/main.dart
@@ -5,7 +5,7 @@
 
 import 'package:artisanal/terminal.dart' show KeyType;
 import 'package:artisanal/tui.dart' as tui;
-import 'package:artisanal_widgets/artisanal_widgets.dart' as w;
+import 'package:artisanal_widgets/widgets.dart' as w;
 
 void main() async {
   final app = w.ArtisanalApp(