+++ pkgs/artisanal_widgets/example/debug_console/main.dart
@@ -1,5 +1,5 @@
 import 'package:artisanal/tui.dart' as tui;
-import 'package:artisanal_widgets/artisanal_widgets.dart' as w;
+import 'package:artisanal_widgets/widgets.dart' as w;
 
 void main() async {
   final controller = w.DebugConsoleController(initiallyVisible: true);