+++ pkgs/artisanal_widgets/example/file_picker/main.dart
@@ -2,7 +2,7 @@ import 'dart:io' as io;
 import 'dart:math' as math;
 
 import 'package:artisanal/tui.dart' as tui;
-import 'package:artisanal_widgets/artisanal_widgets.dart' as w;
+import 'package:artisanal_widgets/widgets.dart' as w;
 
 void main() async {
   await w.runArtisanalApp(