+++ pkgs/artisanal_widgets/example/image/main.dart
@@ -10,7 +10,7 @@ import 'dart:typed_data';
 
 import 'package:artisanal/style.dart' hide Padding, Align;
 import 'package:artisanal/tui.dart' as tui;
-import 'package:artisanal_widgets/artisanal_widgets.dart' as w;
+import 'package:artisanal_widgets/widgets.dart' as w;
 import 'package:image/image.dart' as img;
 
 void main() async {